Move experimental APIs to stable for React 18
See https://github.com/facebook/react/pull/21488.
