+++ types/react-is/index.d.ts
@@ -1,14 +1,10 @@
-// Type definitions for react-is 17.0
+// Type definitions for react-is 18.0
 // Project: https://reactjs.org/
 // Definitions by: Avi Vahl <https://github.com/AviVahl>
 //                 Christian Chown <https://github.com/christianchown>
 //                 Sebastian Silbermann <https://github.com/eps1lon>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 // TypeScript Version: 2.8
-
-// NOTE: Users of the React 18 alpha should add a reference
-// to 'react-is/next' in their project. See next.d.ts's top comment
-// for reference and documentation on how exactly to do it.
 
 export as namespace ReactIs;
 
@@ -33,6 +29,7 @@ export function isProfiler(value: any): value is ReactElement;
 export function isPortal(value: any): value is ReactElement;
 export function isStrictMode(value: any): value is ReactElement;
 export function isSuspense(value: any): value is ReactElement;
+export function isSuspenseList(value: any): value is ReactElement;
 
 export const AsyncMode: symbol;
 export const ContextConsumer: symbol;
@@ -46,3 +43,4 @@ export const Portal: symbol;
 export const Profiler: symbol;
 export const StrictMode: symbol;
 export const Suspense: symbol;
+export const SuspenseList: symbol;
