+++ types/react-is/tsconfig.json
@@ -20,7 +20,6 @@
     },
     "files": [
         "index.d.ts",
-        "test/react-is-tests.tsx",
-        "test/next-tests.tsx"
+        "test/react-is-tests.tsx"
     ]
 }
